App/UserApp/Reservation/requirements.txt
flask
pyjwt
requests
flask_jwt_extended
pytz
gunicorn
azure-servicebus
